improved error handling and simplified running flow
