+++ lib/xcov.rb
@@ -2,7 +2,6 @@ require 'json'
 require 'xcov/version'
 require 'xcov/manager'
 require 'xcov/options'
-require 'xcov/runner'
 require 'xcov/ignore_handler'
 require 'xcov/error_handler'
 require 'xcov/slack_poster'
