+++ lib/xcov/version.rb
@@ -1,6 +1,6 @@
 module Xcov
 
-  VERSION = "1.0.1"
+  VERSION = "1.1.0"
   DESCRIPTION = "xcov is a friendly visualizer for Xcode's code coverage files"
 
 end
